+++ docs/examples/all/plot_report.py
@@ -31,6 +31,7 @@ import tempfile
 from pathlib import Path
 
 from gemseo import configure
+from gemseo.algos.doe.openturns.settings.ot_opt_lhs import OT_OPT_LHS_Settings
 from gemseo.algos.opt.scipy_local.settings.lbfgsb import L_BFGS_B_Settings
 from gemseo.algos.opt.scipy_local.settings.slsqp import SLSQP_Settings
 from gemseo.problems.optimization.rastrigin import Rastrigin
@@ -111,16 +112,14 @@ rastrigin_2d = OptimizationProblemConfiguration(
     "Rastrigin",
     Rastrigin,
     optimum=optimum,
-    doe_size=5,
-    doe_algo_name="OT_OPT_LHS",
+    doe_settings=OT_OPT_LHS_Settings(n_samples=5),
     target_values=target_values,
 )
 rosenbrock_2d = OptimizationProblemConfiguration(
     "Rosenbrock",
     Rosenbrock,
     optimum=optimum,
-    doe_size=5,
-    doe_algo_name="OT_OPT_LHS",
+    doe_settings=OT_OPT_LHS_Settings(n_samples=5),
     target_values=target_values,
 )
 # %%
@@ -145,8 +144,7 @@ rosenbrock_5d = OptimizationProblemConfiguration(
     create_problem,
     target_values=target_values,
     optimum=optimum,
-    doe_size=5,
-    doe_algo_name="OT_OPT_LHS",
+    doe_settings=OT_OPT_LHS_Settings(n_samples=5),
 )
 problems_5d = ProblemsGroup("5D problems", [rosenbrock_5d])
 # %%